ARM64: configs: Make ion allocate in page-sized chunks
With this option disabled, Ion allocates memory in order-4 sized chunks
(16 pages at a time), which is quite a heavy burden on the page
allocator, and can take a while to allocate when memory is fragmented
after a lot of uptime. To lower the allocation latency in Ion, have it
only allocate in page-sized chunks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Joel Gómez <nahuelgomez329@gmail.com>
Signed-off-by: siimsek <malisimsek17@gmail.com>